return to using getBackendRef type function instead of nested for loops (#3)
Introduces `getIndexedBackendRefs` with a similar signature to
`getBackendRefs` but actually returns the backend refs separated by the
index of the rule they belong to. This allows us to only update the
canaryBackendRefs based on that value, and get closer to the original
implementation of this function.

i.e. We remove the need to set the httpRoute through nested for loops
tracking multiple indexes like the following in favor of slices:

```go
httpRoute.Spec.Rules[i].BackendRefs[j].Weight = &desiredWeight
```

becomes:

```go
ref.Weight = &desiredWeight
```